add: seperate dropdown components (user, language) and translate utils
